RF|RenamePage_generics
