+++ src/test/java/school/redrover/model/OrganizationFolderRenamePage.java
@@ -7,36 +7,9 @@ import org.openqa.selenium.support.FindBy;
 import org.openqa.selenium.support.ui.ExpectedConditions;
 import school.redrover.model.base.BasePage;
 
-public class OrganizationFolderRenamePage extends BasePage {
-
-    @FindBy(name = "newName")
-    private WebElement inputNewName;
-
-    @FindBy(name = "Submit")
-    private WebElement renameButton;
-
-    @FindBy(xpath = "//div[@class='warning']")
-    private WebElement warningMessage;
+public class OrganizationFolderRenamePage extends RenamePage {
 
     public OrganizationFolderRenamePage(WebDriver driver) {
         super(driver);
-    }
-
-    public OrganizationFolderRenamePage enterNewName(String name) {
-        inputNewName.clear();
-        inputNewName.sendKeys(name);
-
-        return this;
-    }
-
-    public OrganizationFolderDetailsPage clickRenameButton() {
-        renameButton.click();
-
-        return new OrganizationFolderDetailsPage(getDriver());
-    }
-
-    public String getWarningMessageText() {
-
-        return getWait5().until(ExpectedConditions.visibilityOf(warningMessage)).getText();
     }
 }
